New navigation integration API and also make it optional (#11)
This PR introduces a new API for navigation integration, it is expected
that navigation plugin registers an object using
__register_navigation_plugin global function that has a field mainHook,
which should be a hook that is run inside of the main application
wrapper by the IDE extension.

The hook is expected to return three methods:
1) getCurrentNavigationDexriptor – should return the current navigation
descriptor object
2) requestNavigationChange – a method that takes a navigation
description object and should navigate to a state that corresponds to it

On top of that, the hook is called with an object that has
onNavigationChange method field. This method should be used, whenever
navigation library detects an update in the navigation state (i.e.
someone navigates between screens). The method takes currect navigation
descriptor object as an argument.

The navigation descriptor object is an object that is used to represent
navigation state. Navigation plugin can put any data in that object and
use this data later on when requested to navigate to certain state. The
necessary fields in navigation descriptor are: displayName, and id. The
former is used as a label in url bar in the extension UI while the
latter is used to identify unique routes (e.g. when someone opens screen
A then pushed B and goes back to A, the id 1st and 3rd id's should be
the same as after going back the navigation is the same as at the first
step).

In addition to the API described above this PR introduces some
refactoring:
1) we changed a number of methods and method names that contain
"sztudio" to use prefix "__rnp"
2) we changed runtime such that it only loads a minimum set of
synchronous dependencies and the rest is loaded lazily – this was needed
because after we removed requiring expo-router from runtime the app
stopped working because some of the modules require a specific order of
other modules to be loaded in and that require statement was impacting
this order
3) we extracted expo-router integration to a separate file that is now
loaded only when expo-router files are processed by metro
4) we fixed an issue with strict mode in extension UI that resulted in
initial effect being called twice, as a consequence we'd trigger app
start event twice and would get all processes like
metro/devtools/debugger duplicated