Configure metro to not reset cache (#27)
This PR changes setup for metro to get rid of --reset-cache option
unless for explicit clean builds.

Before, we'd always start metro with reset-cache option thus slowing
down startup. Apparently changing this setting wasn't trivial due to the
fact we've been injecting devtools port number into one of the source
files pulled by the bundle. As a consequence, because of no changes to
that file, metro when using cache would just host the previous version
of it with the old port number.

The main change happens in metro_config where we override
processModuleFilter that uses metro's prelude to assign the port number
at the very beginning of the bundle. The prelude is regenerated with
every run so there is no problem with the port number getting stale
because of cache.

On top of that I made some changes thought the codebase to reverse the
meaning of "use cache" flag to "force build" which in my opinion is more
clear to understand.