Codebase cleanup with numerous changes across the board (#38)
This PR makes a number of changes across the codebase aiming to clean up
some of the problems with its current state:
 - tons of utility functions and files
- duplicated structures with slightly different semantics representing
devices
 - a number of "manager" classes that don't manage too much
- UI is too stateful and the state gets out of sync with the extension
state that keeps the same data

High level overview of what's added and changed here:
1. We introduce API helpers that makes it possible to use interfaces to
define and then call methods from webview. This interface supports
returning data (so you can query things) and also to pass callbacks (so
that you can register listeners). Note that this interface is very
simple and when you want to pass "function" this is only supported as
direct argument (you can't pass structure that has function as one of
the fileds)
2. Rewrite some bits of the codebase to use said API (there are still
many places that can be refactored but they aren't super urgent)
3. Make ExtensionContext available globally as it is only one such
object we receive in activate method. This makes some parts of the code
cleaner as we don't need to pass it around everywhere.
4. Unify device/build representation. Now we only have one "DeviceInfo"
type and one "BuildResult" type
5. Fetch device list directly using simctl and emulator commands to
avoid stored state from getting out of sync
6. Remove ProjectManager as it was mostly just forwarding methods to
Project
7. Some renaming of things from `react-native-preview` to
`react-native-ide`