Expo Dev Client support (#69)
This PR adds support for Expo Dev Client setups. Among some small
changes here are the main things we had to alter:
1) in metro_helper we are setting EXPO_PACKAGER_PROXY_URL env variable.
This variable is used when dev client requests manifest from the
packager. The change is necessary because packager doesn't handle
ephemeral ports well, and when port is set to 0, the manifest includes
links with the default 8081 port. Dev Client uses manifest to get the
bundle URL, which without this change would point to port 8081 instead
of the actual port where metro is running
2) instead of launching the app on iOS with simctl launch, we use expo
deeplink. Otherwise the app launches in the devclient home screen and
requires selecting the server in order to start loading the app. We
retrieve deeplink location by sending request to _expo/link endpoint
which responds with HTTP 307 redirect to the app deeplink URL. Because
iOS shows a dialog when attempting to open deeplink with a given schema
for the first time, we modify scheme approval list config file before
launching the url to prevent it.
3) we also change the way we detect expo, as some projects with
monorepo-type of setup may use expo installed in some top level
directories. This is ok as long as we are able to resolve dependeny on
the expo cli