Report iOS/Android build output using vscode output channel API (#79)
This PR makes it so that long running build processes (xcodebuild for
iOS and gradle for Android) now pass their output as logs to vscode
which allows for better monitoring and diagnosing errors.

Along with this change we made some refactoring to add a utility method
for reading subprocess output line by line. As a part of this PR we are
also updating places in the codebase where we've been using readline
package directly.

In addition, iOS build code has been simplified to focus on options and
parameters we actually use. This makes the code much easier to
understand given we don't need to utilize all the options that were
possible for the time being.