IOS Progress Indication (#124)
Before: 
The only way for a user to get information about build progress was to
analyze logs by hand.
After: 
Progress bar updates according to real or fake build progress.

Implementation: 
Logs generated by xcodebuild (command that we use to execute a build)
contain "Build description path", that points to a directory with
task-store.msgpack file, which contains build task tree. We use that
tree in order to count number of tasks that the build process is going
to execute and then divide no. of completed tasks by that number.

New Dependencies: 
@msgpack/msgpack used to decrypt msgpack file.

Note: 
I also added a "fake progress" that asumes the build is going to take
100 s. It is used when there is any problem with task-store.

---------

Co-authored-by: Filip Andrzej Kaminski <filip@Filips-MacBook-Pro.local>
Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>