Remove logs button from preview loader view (#35)
This PR removes logs button from preview loader screen.

The button was misplaced and caused the whole layout to shift when it'd
appear. After we added log progress, it is no longer needed to provide a
visual indication about the logs being present there. Instead, we make
the progress message clickable.

With this change, the progress message can be clicked at any point. When
build is in progress, we will open build logs panel, and otherwise will
open extension logs.