Move from activity bar to secondary bar hint  (#24)
This PR adds a way to control the panel position.

As a part of this PR we add a new position option "secondary side
panel". We believe it may be useful to use secondary side panel for the
IDE, but it isn't a popular feature of vscode and users are not familiar
with it. On top of that, it cannot be fully controlled programatically.
Specifically, there is no API to move panel from primary to secondary
side panel. To workaround this issue, when user select "secondary side
panel" option we ask them to move to secondary panel by hand using drag
and drop.

This PR also updates the logic for command palette commands and makes it
so that "Open IDE" is always visible unless the panel is opened.

We also do some renamings: Tabpanel -> TabPanel and Sidepanel ->
SidePanel

---------

Co-authored-by: Filip Andrzej Kaminski <filip@Filips-MacBook-Pro.local>
Co-authored-by: Krzysztof Magiera <krzys@swmansion.com>