Fix handling of numeric script ID when debug breakpoint is hit (#46)
This PR fixes issue introduced in
fc3f800806dbe62d5f2e87e456d44c10e8fd2f75 where numeric script IDs were
expected to be numbers. This is not inline with CDP spec, as the IDs are
always strings. As a result we were attempting to parse those numeric
IDs as URLs which resulted in an exception being thrown and debugger
getting stuck without visual indicator.

In this PR I'm also addressing issue with touch events getting triggered
when debug overlay is present. This resulted in touches being sent to
device when someone would click the resume button. As a consequence when
button was placed over an item with click event and breakpoint set, we'd
be stuck in a breakpoint loop.