Use resolved react-native location for locating packager.sh (#74)
This PR updates the way we locate packager.sh script for launching metro
in bare React Native projects.

Before this change, we were assuming `react-native` package is located
in `node_modules` under the main app location. This doesn't need to be
the case for yarn workspaces setups, where `node_modules` are placed in
parent directories of the main application package. We normally take
this into account when loading JS packages and use node's resolution
mechanism, but we weren't doing this for locating packager.sh script as
it isn't a JS package.

This change makes it so that we use node's resolution to locate
`react-native` package location. Then, we find packager script relative
to that location.