Use single message listener for all rpc callbacks (#80)
This PR fixes the issue with message listeners used by rpc callbacks
that were never removed.

We should've remove message listeners when the callback is resolved /
rejected instead of keeping them forever.

This problem amplified issue reported in #69 and fixed in #79

This PR implements a different approach that further optimizes the
number of listeners. Now, instead of registering a listener for every
callback, we keep a map of active callbacks and only register one
message listener as long as there is at least one item in the map. When
given callback promise resolves (gets rejected) we remove the entry from
the map and optionally deregister the listener.