Reveal device preview conditionally (#93)
This PR adds a new behavior that allows for device preview to be shown
despite receiving app ready event.

This is useful for debugging purposes as one of the most common issues
is that the startup process gets stuck on the "waiting for app to load"
message.
In addition, it is also useful for brownfield projects, when app may not
be rendered at launch time, in which case the user won't be able to
navigate to the place in the app when react native is used.

To facilitate the above we:
1) add a click handler on "waiting for app to load" message that'd force
the preview to be displayed. We will document that in the
troubleshooting guide
2) we add new launch configuration based on launch config code
introduced in #90 that disables the promise that waits for the app ready
event

For the above to be possible we also change device session flow a bit,
such that we can deliver preview URL earlier and not only when "ready"
state is reached. This is done by adding another callback that is called
immediately before preview URL is available.