Use unique ids for call and callback ids to avoid duped calls when weboew reloads (#128)
This PR solves an issue where listeners registered from webview would
accumulate over webview reloads.

The problem was due to the fact, we can't unregister listeners when
webview reloads or when it's moved to new window. As a result, listeners
are kept on the extension side and triggered sent to the new context
after webview reloads. Since call and callback ids always start from 1,
we have a collision of ids and those listeners trigger some actions in
the new webview.