Implement jsx transform workaround that doesn't affect order of plugins (#184)
This PR fixes an issue with setups that rely on particular order of JSX
transform plugin relative to other plugins.

Before, we'd add dev version of jsx-transform plugin at the beginning of
the plugins section and disable all other variants of jsx plugin.
However, some setups, specifically apps that use nativewind, rely on a
particular order of jsx transforms for plugin hoisting.

After digging into the code of jsx transform plugin, I noticed, that
across different React Native setups, the dev plugin is already added,
it is just not producing output. The reason for that is that there
JSXElement nodes are transformed by the non-dev version of the plugin
that is typically registered earlier. As a consequence JSXElement
visitor never runs in the dev version.

The workaround this PR implements removes JSXElement visitor from
non-dev version in case the dev version of jsx transform plugin is
present. This is done using a custom plugin that runs before all other
plugins, and we also use `pre` callback to make sure it runs early
enough.