Yet another attempt at solving jsx babel transform overwrites (#216)
Fixes #198 

This is another approach at resolving the complexity of enabling
development JSX transforms.

As noted in #198 the old approach would sometimes produce an error that
the development transform throws when other instance of dev JSX
transform run, or when some of jsx-self or jsx-source transforms are run
on the same file along with the jsx dev version.

The problem in the repro provided in #198 was due to the use of quite
popular react-native-svg-transformer which doesn't forward `plugins`
attribute to the upstream transformer. I proposed a fix for this issue
in that repo
https://github.com/kristerkari/react-native-svg-transformer/pull/354
however, it isn't practical to wait for a fix there and hence the PR
implements another workaround that doesn't depend on the `plugins`
attribute of the transformer.

The new approach relies on additional flag that is turned on when
non-dev version of the plugin is loaded. What we try to achieve is that
only jsx-dev transform is listed as plugin, and other varians (non-dev,
-self, and -source) versions are all turned off. Since we can't hijack
the plugins list, the only option is to use require overwrites.
Furthermore, we need to consider the case when someone has a proper
setup where only dev transform is listed and others are not. If we were
to only overwrite non-deb with dev version, that wouldn't suffice for
this particular case.

What we now do is:
1) disable -self and -source transforms entirely (read inline comment
for details on why we can do it)
2) replace non-dev JSX transform with dev version, and also use a flag
to mark when this replaced non-dev version is used
3) replace dev JSX transform with extra logic that checks for the flag
that marks whether non-dev version is used. If it is used, we return an
empty transform. Otherwise, we just return the dev transform as it means
that non-dev isn't registered.

As stated in the inline comment, this approach would break if someone
has dev transform registered before non-dev. Apparently, I don't see a
better way to move forward with this approach and I'd hope that the
current solution would cover the majority of setups.