fix: fix(lib/expo_router_helpers): avoid unnecessarily closing modals
The fix solves the problem of closing too many modals in the stack. For example, if you had opened four modals in the stack and just wanted to navigate from the fourth modal to the second modal, and this modal was filled with some pieces of information, for instance, in a form, it was all gone. Now it moves to the second modal smoothly, keeping the state in place.