Fix clickability of wrapped items in button-group-top
The button-group-top div contains various buttons, dropdowns, and other interactive elements. When the width of the container becomes small, these items wrap to the next line. However, the wrapped items were not clickable, causing usability issues.
This PR addresses the clickability problem of the wrapped items in the button-group-top div.