Kmagiera/selectable UI (#353)
This PR disables user-select on most of the IDE panel UI. It's done
because now it is easy to accidently select things on the panel
including the simulator image. However, due to the fact we capture mouse
events over simulator and prevent default behavior, it isn't easy to
un-select things. It seems like there's not a lot of value in enabling
user select on the main UI while people may still want to copy things
like error messages from dialogs hence we disable it on the main panel
level and allow it to work on popups and dialogs.