Replace disabled magnet with home button used e.g. for getting back from preview mode. (#366)
When using component preview in non expo-router projects, there is no
way in the panel to go back to the main app from preview mode. Since we
are not currently using the magnet functionality, I decided we could add
"home" button that's similar to what browsers offer, and hence would
provide an intuitive way of going back home from preview mode.

Since it sort-of doubles the reload functionality, we disable it unless
url list is non-empty. This happens specifically when preview is opened
(previews get appended to url list). But home button can also be used in
expo-router projects to reset the stack to /