Fix errors caused by #317 (#367)
In #317 new fields have been added to device settings. However, since we
persist device settings, the changes are not compatible with the old
data that might've been stored there. To prevent this we are adding a
versioning of device settings key (will increment version with future
changes).

This PR also changes the defaults to use disabled location, especially
that we need to have some coordinates specified, it'd be weird to have
everyone using IDE pointing to the same geo location.