Rely on exec extending env by default (#378)
When debugging some weird issues associated with env variables, I
noticed that execa by default extends the process env. In most of the
cases this is the desirable approach and hence we don't need to handle
that ourselves as it only makes the whole flow more polluted with
unnecessary options.