Don't update shared prefs for expo go (#379)
We recently added code that'd write to shaed prefs in order to disable
dev menu popping up on expo dev client on Android. This apparently
doesn't work with expo go, as the go APK installed on the emulator is
not debuggable. In this PR we add a check for the bundle name before we
write the setting.