Handle empty shared preferences file when setting options for Android (#417)
This PR fixes an issue we found with some setups. After the app is
installed for the first time on Android, it sometimes would get the
shared preferences file creates as an empty file. As a result, neither
the parser nor the `cat` command would throw but the resulting prefs
object would be null.

This PR adds a check for the structure of prefs objects, such that we
can be sure it has the expected structure (containing a `map` field in
particular), before we continue with updating the prefs object.

Test:
As I couldn't reproduce this issue on my local environment, the change
that I did was to clear up the shared preferences file after the app was
already installed, then perform a clean build on Android.