Avoid passing PWD entirely to subprocesses when cwd option is set (#432)
This fixes an issue when VSCode started from app launcher would populate
PWD environment constant which interfers with some tooling.

We've already attempted to fix that in #369 but the fix really only
worked when PWD was set as options.

Aparently execa library that we use for subprocess management by default
inheritcs process.env (which is what we want). But in that case env.PWD
would be provided even despite the fact it wasn't included in env from
the options. This got broken along with #378 where we stopped manually
passing process.env when launching.

As a consequence, we will now always override PWD with "0" which will
result in third-party checks to fail on that variable.

Test plan:
Launch react-conf-app repo with VSCode started from launcher

---------

Co-authored-by: Jakub Gonet <jakub.gonet@swmansion.com>