Fix navigation after Preview.  (#499)
This PR fixes a problem with navigation after using preview
functionality.
It turns out that navigation is not ready immediately after mounting the
wrapper component, the solution is to await layout instead.

Bug: 
```
Attempted to navigate before mounting the Root Layout component. Ensure the Root Layout component is rendering a Slot, or other navigator on the first render.
```
To Reproduce: 

go to expo-51 example and try navigating from preview to any path in
main application

Co-authored-by: filip131311 <filip.kaminski@swmansiom.com>