Remove cancel token from fingerprinting, extract fingerprint outside of cache
We need to store fingerprint that was calculated at the start of the build process
instead of calculating it after it finishes, as user can modify workspace while build is happening.

Cancel token was removed because we expect fingerprints to be calculated much quicker than it takes to build the app.