Add build and launch telemetry (#594)
This PR adds more loggind and telemetry around build and launch process.

Building and launching the app is one of the most fragile element of the
IDE pipeline and we get numerous reports around that. While there are a
few known issues around it that we are fixing and investigating. I
figured it'd be good to also add telemetry around that such that we can
track how it works for a larger group of people and also be able to
detect regressions.

## Test plan

Just run the example app with clean build and make sure nothing breaks.