Use metro message endpoint to dispatch dev menu request (#598)
This PR uses a different method for opening dev menu.

Previously we'd call native dev menu module, which was causing problems,
specifically in expo-go where the module resolution was resulting in
bundle error. Since it turned out to be easier to use the standard
method that metro uses for triggering the menu than to debug the old
method, I decided to rework the underlying implementation.

Now, we use metro /message endpoint that is used to proxy messages
between dev tools and device. One of the messages is "devMenu" which
opens the dev menu on the device side (which is what we wanted).

This PR also removes old code which was placed in the wrapper, but also
for android we were using a different method that was relying on adb.
The new method works on both devices, hence we don't need separate
implementation and can delete both of them.

## Test plan
1. Use dev menu option on Android and iOS across all configurations we
have in the repo (expo-go, react-native-75, expo-51 etc)