Use bundle install before pod install when bundle is used (#624)
In #617 we added an option for the IDE to use bundle command instead of
pods directly as it is now recommended in new react native templates and
setups. However, for bundle to work properly, we need to `bundle
install` first. This was missed in that PR as I was testing a project
where bundle install has already run.

This PR changes the pod installation command to: `bundle install &&
bundle exec pod install` as recommended by react native docs. For this
to work we pass `shell` option to the command which allows for joining
the commands.

### How Has This Been Tested: 
1. Delete rn 75 project and checkout a fresh copy
2. Open the fresh RN 75 project with the IDE