Remove JSX code from main runtime.js file (#651)
This PR fixes an issue we detected in some app configurations
(specifically in react conf app). We identified the issue to be rooted
in the fact we've been using JSX syntax in the main runtime.js file. The
runtime is one of the first files included in the bundler. Using JSX
there was making it import react internals, which impacted the order of
the modules being loaded. Apparently, some applications are sensitive to
this order, and the runtime shouldn't impact it.

The fix was to move JSX code inside of wrapper.js that is loaded lazily
and hence doesn't impact the order of modules being processed when
bundle is loaded.

### How Has This Been Tested: 
1. Run react-conf-app example (should successfully start)
2. Add `setWrapperComponentProvider` call to one of the other test apps
to make sure the fix from #616 is still working