Fix setting a 0 0 coordinate on Android   (#681)
This PR fixes an issue with setting android device GPS coordinates. The
command `adb emu geo fix` that we use to set the location has a bug that
causes it to not work when coordinates 0 0 are passed. The fix changes 0
0 to 0.00001 0.00001 instead.

The fix is worthwhile as 0 0 is a natural think to type into the
coordinates box when you want to just check is your application
responses in any way to the change in location. (this is how the bug was
discovered)

### How Has This Been Tested: 

- run android emulator 
- open google maps 
- change location to 0 0