Fix metro cache issue when running project from CLI and IDE interchangeably (#699)
When working on a single project with IDE and from CLI interchangeably,
metro cache may cause bundling issue because of the way we override
files in babel transformer. This can happen when transformed output of
certain files, for which we modify the source code is cached. Although I
couldn't find a reliable way of reproducing this problem, it seem to
randomly pop up when building from CLI and then using the IDE, and can
surface via the CLI bundler or the IDE bundler.

To workaround this problem, we are using cacheVersion metro
configuration option. This lets the IDE to generate its own, unique
cache keys such that to avoid cache collisions between the metro started
via CLI and in the IDE.

Here's a breakdown of changes we're making here:
1) We're adding cacheVersion config option to metro
2) We're adding blockList option to metro's resolver, as fixing the
cache issue revealed another issue with metro resolving files from under
the extension directory when used with expo's fast resolver option
(EXPO_USE_FAST_RESOLVER)
3) We're adding an option to restart the project with metro cache reset
– this way to provide an easy troubleshooting step for the cases when
this or some other metro cache related issue occurs

We're also renaming some env variables starting with REACT_NATIVE_IDE to
RADON_IDE that I discovered in the metro config code.

### How Has This Been Tested: 
1) I used a new create-expo-stack project with tailwind, expo-dev-client
(SDK 51) and EXPO_USE_FAST_RESOLVER option
2) Build the project from CLI and IDE interchangeably several times
until you start getting weird bundle errors
3) After this change, it should no longer happen
4) Test "Clear metro cache" option and see that it restarts the process,
make sure the native build doesn't trigger and also that it takes longer
to bundle first time it loads after this option is selected