Fix breakpoints set in between JS reloads when using new debugger (#786)
This PR fixes the issue when breakpoints set before JS reloads on the
new debugger would be effectively set twice and we wouldn't have a way
of deleting them later on. This issue is a regression introduced in #755

The issue stems from a difference in behavior between old and new
debugger. The old debugger, when JS is reloaded, looses all the
information about the previously set breakpoints. Thanks to that, we can
forget all the breakpoints in our Debug Adapter and rely on VSCode to
request setting them again when new context is instantiated (we send
InitializedRequest).

On the other hand, the new debugger keeps the breakpoints in between
reloads. Therefore, resetting the list of breakpoints in our registry
would lead to the cases where we'd set a single breakpoint for the
second time with CDP after reload. As a consequence, we wouldn't be able
to delete the breakpoint that was set at the beginning.

This PR adds a new parameter that indicates whether the breakpoints get
cleared on context reload. We only set that flag for the old debugger.

### How Has This Been Tested: 
The below is the scenario that was failing:
1. Open app and set breakpoint
2. Reload JS
3. Unset breakpoint
4. Before this change the VM would still be stopping at the breakpoint
that is no longer set with the new debugger.

This scenario was tested in RN 76 and RN 74 projects.