Prevent from sending mouseUp after inspector and right-click inspector (#769)
This PR addresses two issues related to sending "Up" touches to the
sim-server after using the inspector and right-click inspector:
- opening the `InspectorDataMenu` triggered a `mouseLeft` event,
resulting in an "Up" touch
- previously, *only* the `mouseDown` click was ignored after
highlighting an element in inspector mode, while the `mouseUp` event was
still sent

We have now introduced const `shouldPreventFromSendingTouch`, which
prevents sending any mouse events (keyboard events will still be sent)
when the inspector is on.
`resetInspector()` has been moved to the `onMouseUp` event to set
`inspectFrame` to null, replacing its previous placement in the
`mouseDown` event. This change means that inspector mode is exited after
`mouseUp`, preventing the sending of `Up` events.

In onMouseDown `sendInspect` sets inspectFrame after
`onWrapperMouseDown` has already set isPressing to true, and onMouseMove
or onMouseUp then sends a touch (it is caused by throttling). To solve
this, we do not set isPressing to true in `onWrapperMouseDown` for right
mouse clicks.

Fixes #763 

### How Has This Been Tested: 
- Enabled "SIMSERVER_LOG": "debug", then utilized the inspector and
right-click inspector to confirm that "Up" or "Move" touches were no
longer being sent.