Introduce IDE class that manages all necessary class instances (#882)
This PR restructures the logic for creating and accessing some core
extension components like Project, DeviceManager, etc.

Before they were instantiated in webview panel or tabview controller. In
this PR we are introducing a new top level class IDE that controls the
lifecycle of all the listed core components and hooks them up properly
as some instances require other instances to exists.

The main reason for this change is to lay a groundwork for #878 where we
are introducing additional webview panels to be used to display
different tools. The panels require separate webview controller
instances that also need to access some details about project setup that
are currently only kept in the main panel webview controller.

As a side effect of this change, we no longer re-create project when
switching between UI location for the Radon IDE panel (between tab and
side-panel). Therefore this change makes it more seamless for the users
to play with the panel location as updating that no longer restarts the
project.

For the above to work corrently, we add additional logic for handling
`dispose` calls. When the position of the panel is updated, we first get
the original panel disposed, and only then the new panel is created.
When the main one is disposed, we don't know whether that would be
because of the position change or due to the user closing the IDE panel
by hand. In particular, it is possible for user to change the location
w/o using the provided menu options, so there's no indication that we
can have about the fact that the panel location is being updated. In
order to handle that properly, the IDE class has static `attach` method
used by the webview panels and also `detach` method that the panel uses
when they are being disposed. The IDE panel doesn't get disposed
immediately when detached, but it waits for 1 second and if it isn't
attached to a new panel after that time, only then it calls dispose.

### How Has This Been Tested: 
1. Open app in IDE
2. Change panel location between tab/side bar/ new window – notice that
the app doesn't restart when updating the location.
3. Make sure the IDE is functional after the location is changed.
4. Close IDE panel (i.e. just close the tab when panel is displayed as
editor tab) and make sure the IDE is disposed (we print a debug log when
that happens).