Support for launching expo dev plugin tools in separate panels within VSCode (#878)
This PR adds support for redux and react query dev tools expo plugins
and makes their UI display as separate panel in VSCode.

Apart from the support for the listed plugins, this change builds an
infrastructure for supporting more expo dev plugins and other types of
plugins in the future.

Key changes that the PR introduces:
1. We are expanding panels configuration in package.json. Apparently
there's no way to define panels dynamically and we need one panel for
every dev tool such that they can be run side-by-side when necessary. To
avoid having too many panels, every panel is hidden behind its own
context variable that we can dynamically set when we detect the user
requested to open this panel
2. We introduce a change to babel transformer that allows us to capture
the use of expo-dev-plugins we currently support. This way we can detect
which plugins the user has installed and only show relevant tools for
the user to open.
3. For the above to work correctly, we are adding a new type of event
that is passed via the devtools channel (same that we use for events
such as "app ready"). We use this event to notify about the plugins that
the app uses.
4. We're adding a UI component to the top menu bar with a "tools"
dropdown. The dropdown displays the list of available tools and a
checkbox next to every tool allowing the user to enable each tool
separately. When tools is enabled, the new panel displaying the tool UI
will launch.
5. Finally, we're implementing a new "ToolPlugin" interface and its
implementation that coordinates all the expo-dev-plugin specific needs.
The interface allows for tools to be listed as available and implements
a persistent storage for saving which tools are enabled. The
expo-dev-plugin implementation of the tool manages the registration of
webview panels, listening to events about the plugins being used by the
app, and controls the context variables that shows/hide the panel.

As a follow up to this change, we will be adding a docs update that
explains how tools can be used. Currently, we only support
epxo-dev-plugin tools, so for the whole setup to work corrently the user
needs to use Expo and have the specific tools configured according to
the documentation here: https://github.com/expo/dev-plugins

### How Has This Been Tested: 
1. I used a sample Expo project and added redux and react query dev
tools following the docs here: https://github.com/expo/dev-plugins
2. I tested enabling/disabling each plugin separately. When plugin is
enabled we expect the panel to open immediately.
3. I tested resetting metro cache which forces metro to reset at which
point it stars at a new port. This is important as the plugin UI is
hosted via metro. In this scenario is it expected for the panel to close
while the app is reloading.

---------

Co-authored-by: filip131311 <159789821+filip131311@users.noreply.github.com>