Fix debugger after multiple reloads (#913)
This PR adreses a problem with multiple debugSessions being held by
vscode, after performing multiple reloads in quick succession. This was
caused by `debug.stopDebugging(this.vscSession)` not removing the
previous session if called in to rapid succession. We trigger the
`stopDebugging` method is by dispossessing `debugSession`. Because we
dispose of the old `debugSession` every time we create a new one,
attempt to create multiple instances of debug session in a row leads to
the problem.

In our current flow multiple instances of `debugSession` might be
created after multiple restarts happening before, the previous one
finishes. To avoid that we introduce a cancel token to `launchApp`
method (which is triggered by the restart) to exclude parallel
execution.

Fixes #860

### How Has This Been Tested: 

Performed reproduction from #860:
- force bundle error 
- reload the app multiple times 
- fix bundle error 
- the debugger should work properly after this fix