Fix infinite loading when cdp does not return a variable (#928)
This PR fixes a problem with debugger infinitely awaiting for variable
from cdp, blocking the whole debugger process in consequence. The
solution is to add a time out to the function that fetches the
properties and return empty variable list if it passes.

### How Has This Been Tested: 

- add a `console.log(devtoolsAgent)` somewhere in `RNIDE_inspect`
section of wrapper.js
- click through returned object before the change the
`_persistedSelection` filled would load infinitely after it stops
awaiting after 3 s.