First-party redux support (#931)
This PR adds first-party support for redux. I resued the enhancer part
and WebUI from
https://github.com/matt-oakes/redux-devtools-expo-dev-plugin - the code
seems to do everything we need with nicely separated communication,
except that the communication through the `expo/devtools` is hardcode. I
created a fork of it that's included as a submodule in our repo, but I'm
trying to upstream the changes with communication logic abstraction
(https://github.com/matt-oakes/redux-devtools-expo-dev-plugin/pull/19)
so we can easily keep the fork up to date. The rest (communication
through vscode) can be found in
https://github.com/software-mansion-labs/redux-devtools-expo-dev-plugin/pull/2).

To archive zero-config support, we utilize
`window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__`, where we pass compose
from `redux-devtools-expo-dev-plugin` with our communication injected.
Redux dev tools communicate with the extension through the dev tools,
and the extension passes messages through webview events to WebUI.

We need to have `redux-devtools-expo-dev-plugin` built to import through
a wrapper in app runtime. Similarly, the webui needs to be bundled, thus
I added `packages/vscode-extension/scripts/install-dev-plugins.sh`,
currently the product of it is committed, and it needs to be run
manually, but this is subjected to change.

### How Has This Been Tested: 

- Open `expo-52-prebuild-with-plugins`
- Comment out configuration related to expo dev plugins
```
const store = configureStore({
  reducer: rootReducer,
  // devTools: false,
  // enhancers: (getDefaultEnhancers) =>
  //   getDefaultEnhancers().concat(devToolsEnhancer()),
});
```
- Reload the app
- Confirm redux dev tools work