Fix startup theme glitch (#940)
This PR improves upon  #938. 

Instead of the CSS trick for empty attribute and holding the rendering,
we just provide the **proper** initial value in the html. This seems to
be the smoother solution.

I'm using `workspace.getConfiguration("RadonIDE");` instead of
`IDE.getInstance` as IDE is not attached at this point yet.

### How Has This Been Tested: 

- Restart the extension and see if the glitch exists