Cleanup dev plugins architecture along with redux plugin (#943)
In #931 we introduces a buil-in support for redux plugin that doesn't
rely on expo dev plugins package. This PR however brought a few
potential issues to the codebase, so here we're refactoring it aiming to
fix those problems while also preparing for more steamlined support for
other tools:

1) removing console log statements introduced in #931 that were
polluting the app output
2) clearing up redux registration such that it only relies on the
"compose" callback and doesn't require subsequent calls for setting
devtools agent nor hook-based integration. This is done by accessing the
devtools agent in the same way we do that in the wrapper component.
3) getting rid of direct imports of devtools related code – this for
unknown reason turns out to be causing issues with metro as direct
imports would impact the order in which modules are loaded and hence may
impact the app behavior.
4) we're refactoring module registration such that it is no longer
specific to expo. Before we had registry for expo-related plugins and
now in #931 a separate registration process was introduced for redux
plugin specifically. We're making it more generic and porting both expo
plugins and the redux plugin to use it.

### How Has This Been Tested: 
1. Fire up example app with plugins
2. Test redux plugin shows up on the list when redux is used and doesn't
show up otherwise.
3. Test scenario from #931