Adding network tools panel (#933)
This PR adds new tools panel displaying XHR network traffic.

The panel UI is based off of the chrome devtools interface. We prepare a
special build that only shows the networking panel with the necessary
modifications done here:
https://github.com/software-mansion-labs/chrome-devtools-ui

The new plugin works as follows:
- when enabled, we use XHR API to hijack HXR object callbacks, we use
the data from callbacks to send information about the requests. The
messages are sent according to the CDP spec and formatted the right way
such that the CDP frontend can understand them w/o any issues
- we spin up a new websocket server which acts as the CDP host. This
host is used to deliver the CDP information about network events to the
frontend.
- when panel is open, we load chrome devtools ui that points to out CDP
websocket server
- finally, the panel reads the messages from our server and displays
them accordingly.

### How Has This Been Tested: 
- Added fetch call to one of the test apps and verify the requests show
up in the network inspector
 - Test scenarios when panel is disabled/enabled