Add alternative implementation for WeakRefs for RN versions that doesn't support WeakRefs (#951)
In #933 we added network devtools panel integration. The implementation
relied on WeakRefs to keep references to XHR objects that can later be
queried by the network panel. However, WeakRefs aren't supported across
all versions/configurations in RN.

This PR introduces an alternate implementation that clears up the
references to XHR after 3minutes. This is needed such that we don't keep
all the network requests in memory forever. We also make a change here
that removes the kept object as soon as the details about the request
are fetched. This will allow us to cleanup the refs as soon as the UI
panel fetches the request body details.

### How Has This Been Tested: 
1. Test fetch with expo-51 and expo-52. Add console.log to make sure the
alternate implementation is only used in the former.
2. Before this change, fetches wouldn't work at all when network panel
was enabled.