Add scrolling support (#957)
This PR adds support for scrolling in Radon IDE.

The scroll is simulated by simulator-server and reflected as a swipe
(the direction of the swipe is opposite to wheel direction). The
captured `WheelEvent` deltas are normalized to provide three predefined
scroll magnitudes (short, medium and long) and negated before passing to
simulator server for better effect (passing raw values makes it hard to
simulate a short scroll).


https://github.com/user-attachments/assets/b49dcf90-7ea0-4348-965a-cf1ebec0cb4e


https://github.com/user-attachments/assets/59fd7a74-c555-4ba3-b482-a96e8a68d187