Add node version check (#971)
This PR adds a node version check to the start up flow that checks node
version found in the PATH to the minimum version specified either by the
package.json or react-native dependency package.json.

### How Has This Been Tested: 

-run `react-native-78` test app and change minimum versions in
package.json and in react-native dependency, check if the warning
appears in "run diagnostics screen".

---------

Co-authored-by: Juliusz Wajgelt <49338439+jwajgelt@users.noreply.github.com>