"Outline Renders" tool (#967)
Adds the "Outline Renders" tool which highlights React components when
they are rendered. Based on
[react-scan](https://github.com/aidenybai/react-scan) utility.

- we reuse the [`react-scan`](https://github.com/aidenybai/react-scan)
implementation of drawing the outlines on the canvas. Accessing it from
our frontend code required abstracting the renderer code and exporting
it from the package. These changes are available and pulled from a
public fork the `react-scan` repository.
https://github.com/software-mansion-labs/react-scan/pull/1
- we use the [`bippy`](https://github.com/aidenybai/bippy) library for
instrumenting the React rendering process. We include the built library
code (extracted from the official release) in this repo in the
`packages/vscode-extension/lib/bippy` directory and inject it into the
RN Application bundle.
We do it this way because it needs to be included in the final extension
package in order to be injected into the RN application, and only the
`lib` directory is available to Metro.

### How Has This Been Tested: 
- Open the "Tools" dropdown and toggle "Outline Renders".
- Trigger component updates in the application.
- The updated components should be briefly highlighted, with their names
and update counter displayed.
- Resizing the phone Preview should not break the outlines' positions.
- The tool's enabled status should persist app, extension and IDE
launches/reloads.