Add a fallback path to the end of the path variable (#995)
This PR fixes the issue introduced by #975, that was causing for the
path variable to be incomplete.
The problem could be surfaced when trying to use `sim-server
fingerprint` functionality as it relies on the PATH variable.

The solution is to in addition to the new path finding method, append
the old path at the end of the new one to fill in any missing paths.

### How Has This Been Tested: 

- Run `react-native-78` test app
- add license