Add first-party react-query devtools support (#973)
This PR adds first-party react-query dev tools support. It works without
any additional configuration and does not conflict with the Expo Dev
plugin, so the old implementation, which is based on Expo Dev Plugins
(#878), is being removed.

**The solution works as follows:** 
- While transforming the react-query library, we inject a code that
calls `window.__RNIDE_REACT_QUERY_CLIENT_INIT__` with `queryClient` on
mount;
- `__RNIDE_REACT_QUERY_CLIENT_INIT__` broadcasts the apps react-query
state using RN dev tools;
- Radon IDE extension is used as a proxy - receives the events from RN
dev-tools and sends them to React Query DevTools webview;
- Thanks to the above, webview has synced react-query state and displays
it using [original react-query
devtools](https://github.com/TanStack/query/tree/main/packages/query-devtools).

**WebUI**

WebUI needs to be bundled so it can be inserted into webview. For that,
I created a separate repo
https://github.com/software-mansion-labs/radon-ide-react-query-webui -
which consists of
https://github.com/TanStack/query/tree/main/packages/query-devtools with
additional logic required for syncing the state between app and dev
tools. Built artifacts of webui are committed to the webui repo and are
copied while building the Radon IDE extension.

**Communication**

The reason is to sync both states (apps with dev tools one). The sync is
based on
https://github.com/TanStack/query/blob/main/packages/query-broadcast-client-experimental/src/index.ts,
but instead of `BroadcastChannel`, we communicate through RN dev-tools
similarly to redux dev tools #931.

### How Has This Been Tested: 

- run `eas-52-prebuild-with-plugins`
- check if `React Query DevTools`, are available in tools list 
- make sure `React Query DevTools DevPlugin` is not visible
- open `React Query DevTools`
- go to plugins page
- press React Query counter 
- confirm that the events are visible